Merging with master

+++ src/evaluator.cc
@@ -89,7 +89,7 @@ Evaluator::Outcome::Outcome( const AnswerBuffers::Outcome & dna )
 }
 
 Evaluator::Outcome Evaluator::score( WhiskerTree & run_whiskers,
-				     const bool trace, const unsigned int carefulness ) const
+				     const bool trace, const double carefulness ) const
 {
   return score( run_whiskers, _prng_seed, _configs, trace, TICK_COUNT * carefulness );
 }
